Rewrite start-game usecase
